+++ std-bits/table/src/main/java/org/enso/table/aggregations/MinOrMax.java
@@ -6,6 +6,7 @@ import org.enso.base.ObjectComparator;
 import org.enso.table.data.column.storage.Storage;
 import org.enso.table.data.table.Column;
 import org.enso.table.data.table.problems.InvalidAggregation;
+import org.graalvm.polyglot.Context;
 
 /**
  * Aggregate Column finding the minimum (minOrMax = -1) or maximum (minOrMax = 1) entry in a group.
@@ -32,6 +33,7 @@ public class MinOrMax extends Aggregator {
 
   @Override
   public Object aggregate(List<Integer> indexes) {
+    Context context = Context.getCurrent();
     Object current = null;
     for (int row : indexes) {
       Object value = storage.getItemBoxed(row);
@@ -50,6 +52,8 @@ public class MinOrMax extends Aggregator {
           return null;
         }
       }
+
+      context.safepoint();
     }
     return current;
   }